+++ tests.py
@@ -1,6 +1,6 @@
-import unittest
+import unittest, os, sys
 
-import ZeltInit
+import ZeltInit, config
 
 Z = ZeltInit
 
@@ -14,6 +14,8 @@ def simulate_round(turns):
 class MyTest(unittest.TestCase):
     def setUp(self):
         print("*****")
+        config.TrackerConfig(os.path.dirname(sys.executable))
+        # print(Z.config)
 
         Z.set_up_test()
 
@@ -40,8 +42,6 @@ class MyTest(unittest.TestCase):
         # (tricks, att_trick, def_trick))
         i = 1
 
-
-
         known_values = (
             (1, 4, -1, False, True, True, (False, 0, 0)),
             (5, 5, -9, False, True, True, (False, 0, 0)),
@@ -66,15 +66,13 @@ class MyTest(unittest.TestCase):
 
             Z.handle_withering(combatants, damage, tricks)
 
-
-
             a.shift_target = None
             d.shift_target = None
 
             self.assertEqual(a.initiative, n_attacker_init,
                              "loop " + str(
                                      i) + " " + a.name + " attacking " + d.name + ". Initiative is " + str(
-                                 a.initiative) + " but should be " + str(n_attacker_init))
+                                     a.initiative) + " but should be " + str(n_attacker_init))
             self.assertEqual(a.crash_state, a_crash,
                              str(a) + "'s a_crash status is not correct.")
 
@@ -86,7 +84,6 @@ class MyTest(unittest.TestCase):
             # Trick Assertions
             i += 1
             Z.sort_table()
-
 
     def test_initiative_shift(self):
         print("Testing for Shift")
@@ -120,14 +117,14 @@ class MyTest(unittest.TestCase):
             a, d = Z.character_list[attacker], Z.character_list[defender]
 
             print("")
-            Z.print_table()
+
             print("")
 
             Z.handle_withering(combatants, damage, trick)
             print(attacker, "attacks", defender)
 
             print("")
-            Z.print_table()
+
             print("")
 
             self.assertEqual(d.initiative, n_def_init)
@@ -212,10 +209,21 @@ class MyTest(unittest.TestCase):
         self.assertEqual(a.initiative, n_att_init)
         self.assertTrue(a.crash_state)
         self.assertEqual(a.crash_counter, counter)
+
+        #  Get arnold back to the top of the initiative
+        Z.skip_turn()
+        Z.sort_table()
+        Z.skip_turn()
+        Z.sort_table()
+        Z.skip_turn()
+        Z.sort_table()
+        Z.skip_turn()
+        Z.sort_table()
+
         combatants, damage, n_att_init, n_def_init, crash_status, counter = next(known_values)
         a, d = Z.character_list[combatants[0]], Z.character_list[combatants[1]]
         Z.handle_withering(combatants, damage)
-        Z.sort_table()
+
         self.assertEqual(a.initiative, n_att_init)
         self.assertFalse(a.crash_state)
         self.assertEqual(a.crash_counter, counter)
@@ -251,7 +259,7 @@ class MyTest(unittest.TestCase):
             ((4, 0), 0, -6, 1, True, 2),
             ((4, 0), 0, -5, 1, True, 3),
             ((4, 0), 0, -5, 1, True, 3),
-            ((4, 0), 1, 5, 17, True, 0),
+            ((0, 1), 1, 5, 20, True, 0),
         )
 
         for value in values:
@@ -297,6 +305,121 @@ class MyTest(unittest.TestCase):
 
         for character in removed_from_play:
             self.assertNotIn(character, Z.character_list)
+
+    def test_onslaught_penalty(self):
+        Z.character_list = []
+        Z.add_npc("Amber", False, 0, initiative=10)
+        Z.add_npc("Billy", False, 0, initiative=8)
+        Z.add_npc("Carol", False, 0, initiative=6)
+        Z.add_npc("Danny", False, 0, initiative=4)
+        Z.sort_table()
+
+        values = (
+            (0, 3, -1),  # (Attacker, Defender, new Onslaught
+            (0, 2, -2),
+            (0, 1, -3),
+            (0, 2, -1),
+
+        )
+
+        self.assertEqual(Z.character_list[0].onslaught, 0)
+        for attacker, defender, onslaught in values:
+            # print
+            combatants = attacker, defender
+            a = Z.character_list[attacker]
+            d = Z.character_list[defender]
+            Z.handle_withering(combatants, 0)
+            self.assertEqual(d.onslaught, onslaught, str(d.name) + "'s onslaught should be " + str(onslaught))
+            self.assertEqual(a.onslaught, 0)
+            Z.sort_table()
+            print("")
+
+        Z.character_list = []
+        Z.add_npc("Amber", False, 0, initiative=10)
+        Z.add_npc("Billy", False, 0, initiative=8)
+        Z.add_npc("Carol", False, 0, initiative=6)
+        Z.add_npc("Danny", False, 0, initiative=4)
+        Z.sort_table()
+        for attacker, defender, onslaught in values:
+            # print
+            combatants = attacker, defender
+            a = Z.character_list[attacker]
+            d = Z.character_list[defender]
+            Z.handle_decisive(combatants, False)
+            self.assertEqual(d.onslaught, onslaught, str(d.name) + "'s onslaught should be " + str(onslaught))
+            self.assertEqual(a.onslaught, 0)
+            Z.sort_table()
+            print("")
+
+        Z.character_list = []
+        Z.add_npc("Amber", False, 0, initiative=10)
+        Z.add_npc("Billy", False, 0, initiative=8)
+        Z.add_npc("Carol", False, 0, initiative=6)
+        Z.add_npc("Danny", False, 0, initiative=4)
+        Z.sort_table()
+        for attacker, defender, onslaught in values:
+            # print
+            combatants = attacker, defender
+            a = Z.character_list[attacker]
+            d = Z.character_list[defender]
+            Z.handle_gambits(combatants, False, "Disarm")
+            self.assertEqual(d.onslaught, onslaught, str(d.name) + "'s onslaught should be " + str(onslaught))
+            self.assertEqual(a.onslaught, 0)
+            Z.sort_table()
+            print("")
+
+    def test_legendary_size(self):
+        Z.reset_combat()
+        jack = Z.Character(name='Jack', initiative=3)
+        Z.character_list.append(jack)
+        giant = Z.Character(name='Giant', initiative=10, legendary_size=True)
+        Z.character_list.append(giant)
+        Z.sort_table()
+        i = 0
+
+        known_values = (
+            (5, 9, 5, False),
+            (7, 17, 1, False),
+            (10, 33, -9, True),
+            (1, 35, -9, True),
+            (10, 46, -19, True)
+        )
+
+        for damage, j_init, g_init, crash in known_values:
+            if Z.character_list[0] != jack:
+                Z.skip_turn()
+                Z.sort_table()
+            Z.print_table(True)
+            Z.handle_withering((0, 1), damage)
+            Z.print_table(True)
+            self.assertEqual(giant.initiative, g_init, "loop " + str(i))
+            self.assertEqual(jack.initiative, j_init, "loop " + str(i))
+            self.assertEqual(giant.crash_state, crash, "loop " + str(i))
+            i += 1
+
+            # values = iter(known_values)
+            #
+            # damage, j_init, g_init, crash = next(values)
+
+            # Z.skip_turn()
+            # Z.sort_table()
+            # Z.handle_withering((0, 1), damage)
+            # Z.sort_table()
+
+            # Z.print_table(True)
+            # self.assertEqual(giant.initiative, g_init)
+            # self.assertEqual(jack.initiative, j_init)
+            # self.assertEqual(giant.crash_state, crash)
+            #
+            # damage, j_init, g_init, crash = next(values)
+            # Z.sort_table()
+            # Z.handle_withering((0, 1), damage)
+            # Z.sort_table()
+            #
+            # Z.print_table(True)
+            # self.assertEqual(giant.initiative, g_init)
+            # self.assertEqual(jack.initiative, j_init)
+            # self.assertEqual(giant.crash_state, crash)
 
 
 if __name__ == '__main__':